backend/requirements.txt
fastapi
uvicorn
pymongo
geojson
python-dotenv